casting to float before generating numpy array

+++ DeepSeekMoE_simplification_service.py
@@ -33,9 +33,9 @@ class DeepSeekMoE_Simplification_Service(MoEITS_Simplification_Service):
             for j, e2 in enumerate(experts):
                 if i<j:
                     print(experts[i].gate_proj.weight)
-                    gate_info = compute_information_measures(experts[i].gate_proj.weight.detach().numpy(), experts[j].gate_proj.weight.detach().numpy())['NMI']
-                    up_info = compute_information_measures(experts[i].up_proj.weight.detach().numpy(), experts[j].up_proj.weight.detach().numpy())['NMI']
-                    down_info = compute_information_measures(experts[i].down_proj.weight.detach().numpy(), experts[j].down_proj.weight.detach().numpy())['NMI']
+                    gate_info = compute_information_measures(experts[i].gate_proj.weight.detach().float().numpy(), experts[j].gate_proj.weight.detach().float().numpy())['NMI']
+                    up_info = compute_information_measures(experts[i].up_proj.weight.detach().float().numpy(), experts[j].up_proj.weight.detach().float().numpy())['NMI']
+                    down_info = compute_information_measures(experts[i].down_proj.weight.detach().float().numpy(), experts[j].down_proj.weight.detach().float().numpy())['NMI']
                     results[i,j] = (gate_info+up_info+down_info)/3
                     results[j,i] = (gate_info+up_info+down_info)/3
